pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etsy-unified-os"
version = "0.1.0"
description = "Unifies Etsy skills/projects from local sources and GitHub into one operational repo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "PyYAML>=6.0",
  "requests>=2.32",
  "python-dotenv>=1.0"
]

[project.scripts]
etsy-os = "etsy_unified_os.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
